Rename go -> just-next-tick
